Add dependency on backports.weakref
One of our dependencies (backports.tempfile) is dependent on
backports.weakref, which doesn't have any non-pre-release versions. To
avoid users having to specify the --pre flag when installing floyd-cli
with some versions, we will explicitly call out the pre-release
dependency on backports.weakref. When 1.0 is released, we can remove
this explicit declaration for backports.weakref>=1.0rc1